+++ src/exml/Exml.ts
@@ -11,12 +11,13 @@ import { devlog } from "../helper";
 export default class Exml extends Listener {
 	public constructor(protected subscriptions: vscode.Disposable[]) {
 		super();
+		devlog(this,"constructor");
 		this.addListener(vscode.languages.registerHoverProvider(['typescript'], new ExmlHoverProvider()));
 		this.addListener(vscode.languages.registerDefinitionProvider(['typescript'], new ExmlLinkProvider()));
 		this.addListener(vscode.languages.registerCompletionItemProvider(['typescript'], new ExmlPathAutoCompleteProvider(), "="));
 
 		this.addListener(vscode.commands.registerCommand("egret-helper.goToExml", () => {
-			devlog("Exml constructor egret-helper.goToExml");
+			devlog(this,"egret-helper.goToExml");
 			new Runner().exec();
 		}));
 	}
@@ -26,7 +27,7 @@ class Runner {
 	public exec() {
 		let activieWin = vscode.window.activeTextEditor;
 		if (!activieWin) {
-			devlog("未找到激活的窗口")
+			devlog(this,"未找到激活的窗口")
 			return;
 		}
 		let doc = activieWin.document;